+++ Examples/Controls Example/SwipeViewExample/ViewController.m
@@ -11,10 +11,6 @@
 
 @implementation ViewController
 
-@synthesize swipeView = _swipeView;
-@synthesize label = _label;
-
-
 - (void)viewDidLoad
 {
     [super viewDidLoad];
@@ -22,7 +18,6 @@
     //configure swipe view
     _swipeView.alignment = SwipeViewAlignmentCenter;
     _swipeView.pagingEnabled = YES;
-    _swipeView.wrapEnabled = NO;
     _swipeView.itemsPerPage = 1;
     _swipeView.truncateFinalPage = YES;
 }
@@ -31,10 +26,6 @@
 {
     _swipeView.delegate = nil;
     _swipeView.dataSource = nil;
-    
-    [_swipeView release];
-    [_label release];
-    [super dealloc];
 }
 
 - (BOOL)shouldAutorotateToInterfaceOrientation:(UIInterfaceOrientation)interfaceOrientation
@@ -48,7 +39,7 @@
     //normally we'd use a backing array
     //as shown in the basic iOS example
     //but for this example we haven't bothered
-    return 10;
+    return 100;
 }
 
 - (UIView *)swipeView:(SwipeView *)swipeView viewForItemAtIndex:(NSInteger)index reusingView:(UIView *)view
@@ -59,7 +50,7 @@
         //control events are bound to view controller in nib file
         //note that it is only safe to use the reusingView if we return the same nib for each
         //item view, if different items have different contents, ignore the reusingView value
-    	view = [[[NSBundle mainBundle] loadNibNamed:@"ItemView" owner:self options:nil] lastObject];
+    	view = [[NSBundle mainBundle] loadNibNamed:@"ItemView" owner:self options:nil][0];
     }
     return view;
 }